+++ src/app/beer_garden/db/mongo/models.py
@@ -353,7 +353,6 @@ class Request(MongoModel, Document):
     command_type = StringField(choices=BrewtilsCommand.COMMAND_TYPES)
     created_at = DateTimeField(default=datetime.datetime.utcnow, required=True)
     updated_at = DateTimeField(default=None, required=True)
-    expiration_at = DateTimeField(default=None, required=False)
     status_updated_at = DateTimeField()
     error_class = StringField(required=False)
     has_parent = BooleanField(required=False)
@@ -545,32 +544,6 @@ class Request(MongoModel, Document):
                     self.has_parent = False
                     self.parent = None
 
-        if not self.expiration_at and self.status in BrewtilsRequest.COMPLETED_STATUSES:
-            # If parent or orphaned
-            if not self.has_parent or self.command_type in ["TEMP", "ADMIN"]:
-                if self.command_type == "INFO":
-                    ttl = config.get("db.prune.ttl.info", default=-1)
-                    if ttl > -1:
-                        self.expiration_at = self.created_at + datetime.timedelta(
-                            minutes=ttl
-                        )
-                elif self.command_type == "ACTION":
-                    ttl = config.get("db.prune.ttl.action", default=-1)
-                    if ttl > -1:
-                        self.expiration_at = self.created_at + datetime.timedelta(
-                            minutes=ttl
-                        )
-                else:
-                    # TEMP or ADMIN
-                    self.expiration_at = datetime.datetime.utcnow()
-
-            if self.has_parent and self.parent is not None and not self.expiration_at:
-                parent = Request.objects(id=self.parent.id).only("expiration_at")
-                if parent:
-                    expiration_at = getattr(parent, "expiration_at", None)
-                    if expiration_at:
-                        self.expiration_at = expiration_at
-
         if not self.has_parent or self.parent is None:
             if not self.root_command_type:
                 self.root_command_type = self.command_type
@@ -588,24 +561,10 @@ class Request(MongoModel, Document):
                 # to the same as this request
                 self.root_command_type = self.command_type
 
-    def _set_child_expiration(self):
-
-        updates = Request.objects(parent=self, expiration_at=None).update(
-            set__expiration_at=self.expiration_at
-        )
-        if updates > 0:
-            for child_request in Request.objects(parent=self).only("expiration_at"):
-                child_request._set_child_expiration()
-
     def _post_save(self):
 
         if self.status == "CREATED":
             self._update_raw_file_references()
-
-        if (
-            self.expiration_at or not self.has_parent
-        ) and self.status in BrewtilsRequest.COMPLETED_STATUSES:
-            self._set_child_expiration()
 
     def _update_raw_file_references(self):
         parameters = self.parameters or {}